requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
flask>=3.0.0
flask-cors>=4.0.0
openpyxl>=3.1.0
